requirements.txt
a2wsgi==1.10.7
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
altair==5.4.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bentoml==1.3.9
billiard==4.2.1
blinker==1.8.2
cachetools==5.5.0
cattrs==23.1.2
cbor==1.0.0
celery==5.3.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==44.0.2
cssselect==1.2.0
databases==0.9.0
datasets==2.19.0
deepmerge==2.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dulwich==0.22.3
ecdsa==0.19.0
elastic-transport==8.17.1
elasticsearch==8.17.2
environs==9.5.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.16.1
FlagEmbedding==1.3.2
flatbuffers==24.3.25
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-ws==0.6.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
ijson==3.3.0
importlib-metadata==6.11.0
inflection==0.5.1
inscriptis==2.5.0
ir_datasets==0.5.8
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
loguru==0.7.3
lxml==5.3.0
lz4==4.3.3
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.23.1
mdurl==0.1.2
milvus-model==0.2.8
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.9.3
networkx==3.4.1
numpy==1.26.4
nvidia-ml-py==11.525.150
ollama==0.1.8
openai==1.51.2
openllm==0.6.13
openpyxl==3.1.5
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
peft==0.13.2
pillow==11.1.0
pip-requirements-parser==32.0.1
pipx==1.7.1
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pyaml==24.9.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymilvus==2.4.9
PyMySQL==1.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
questionary==2.0.1
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
simple-di==0.1.5
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
thulac==0.2.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
tornado==6.4.1
tqdm==4.66.5
trec-car-tools==2.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unlzw3==0.2.2
urllib3==2.2.3
userpath==1.9.2
uv==0.4.25
uvicorn==0.32.0
vaderSentiment==3.3.2
vine==5.1.0
w3lib==2.2.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==0.24.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
zlib-state==0.1.9
